+++ src/main/java/GPT/Command.java
@@ -1,3 +1,5 @@
+package GPT;
+
 public abstract class Command {
     public abstract void execute(TaskList taskList, Ui ui, Storage storage);
     public boolean isExit() {
