+++ src/main/java/GPT/Deadline.java
@@ -1,5 +1,7 @@
+package GPT;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+
 
 class Deadline extends Task {
     private LocalDateTime by;
