+++ src/main/java/GPT/DeleteCommand.java
@@ -1,16 +1,18 @@
-public class MarkCommand extends Command {
+package GPT;
+public class DeleteCommand extends Command {
     private int index;
 
-    public MarkCommand(String input) {
+    public DeleteCommand(String input) {
         this.index = Integer.parseInt(input.split(" ")[1]) - 1;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         try {
-            taskList.markTask(index);
+            Task removedTask = taskList.getTasks().get(index);
+            taskList.deleteTask(index);
             storage.saveTasks(taskList.getTasks());
-            ui.showTaskAdded(taskList.getTasks().get(index), taskList.getTasks().size());
+            ui.showTaskRemoved(removedTask, taskList.getTasks().size());
         } catch (GPTException e) {
             ui.showError(e.getMessage());
         }
