+++ src/main/java/GPT/Event.java
@@ -1,3 +1,4 @@
+package GPT;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
