+++ src/main/java/GPT/EventCommand.java
@@ -1,3 +1,4 @@
+package GPT;
 import java.time.LocalDateTime;
 
 public class EventCommand extends Command {
