+++ src/main/java/GPT/ExitCommand.java
@@ -1,3 +1,4 @@
+package GPT;
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
