+++ src/main/java/GPT/GPTException.java
@@ -1,3 +1,4 @@
+package GPT;
 public class GPTException extends Exception {
     public GPTException(String message) {
         super(message);
