+++ src/main/java/GPT/MarkCommand.java
@@ -1,17 +1,17 @@
-public class DeleteCommand extends Command {
+package GPT;
+public class MarkCommand extends Command {
     private int index;
 
-    public DeleteCommand(String input) {
+    public MarkCommand(String input) {
         this.index = Integer.parseInt(input.split(" ")[1]) - 1;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         try {
-            Task removedTask = taskList.getTasks().get(index);
-            taskList.deleteTask(index);
+            taskList.markTask(index);
             storage.saveTasks(taskList.getTasks());
-            ui.showTaskRemoved(removedTask, taskList.getTasks().size());
+            ui.showTaskAdded(taskList.getTasks().get(index), taskList.getTasks().size());
         } catch (GPTException e) {
             ui.showError(e.getMessage());
         }
