+++ src/main/java/GPT/Task.java
@@ -1,3 +1,4 @@
+package GPT;
 abstract class Task {
     protected String description;
     protected boolean isDone;
