+++ src/main/java/GPT/TaskType.java
@@ -1,3 +1,4 @@
+package GPT;
 public enum TaskType {
     TODO,
     DEADLINE,
