+++ src/main/java/GPT/Todo.java
@@ -1,3 +1,4 @@
+package GPT;
 class ToDo extends Task {
     public ToDo(String description) {
         super(description, TaskType.TODO);
