+++ src/main/java/GPT/TodoCommand.java
@@ -1,3 +1,4 @@
+package GPT;
 public class TodoCommand extends Command {
     private String description;
