+++ src/main/java/GPT/Ui.java
@@ -1,3 +1,4 @@
+package GPT;
 import java.util.ArrayList;
 import java.util.Scanner;
