+++ src/main/java/GPT/UnmarkCommand.java
@@ -1,3 +1,4 @@
+package GPT;
 public class UnmarkCommand extends Command {
     private int index;
